+++ RELEASES.md
@@ -8,6 +8,7 @@ Released on TBD (UTC).
 - Added Renko bar aggregator (#2941), thanks @faysou
 - Added `time_range_generator` for on the fly data data subscriptions (#2952), thanks @faysou
 - Added `__repr__` to `NewsEvent` (#2958), thanks @MK27MK
+- Added `convert_quote_qty_to_base` config option to `ExecEngineConfig` (default `True` to retain current behavior) allows adapters to keep quote-denominated sizes when needed
 - Added contingent order fields `parent_order_id` and `linked_order_ids` for `OrderStatusReport` and reconciliation
 - Added Binance BBO `price_match` parameter support for order submission
 - Added BitMEX conditional orders support
@@ -80,7 +81,7 @@ None
 - Fixed some broken links
 
 ### Deprecations
-None
+- Deprecated `convert_quote_qty_to_base`; disable (`False`) to maintain consistent behaviour going forwards. Automatic conversion will be removed in a future version.
 
 ---